+++ src/ngio/common/_roi.py
@@ -11,9 +11,7 @@ import numpy as np
 import zarr
 from pydantic import BaseModel, ConfigDict, Field
 
-from ngio.common._array_io_pipe import (
-    SlicingInputType,
-    TransformProtocol,
+from ngio.common._array_io_pipes import (
     build_dask_getter,
     build_dask_setter,
     build_masked_dask_getter,
@@ -23,6 +21,7 @@ from ngio.common._array_io_pipe import (
     build_numpy_getter,
     build_numpy_setter,
 )
+from ngio.common._array_io_utils import SlicingInputType, TransformProtocol
 from ngio.common._dimensions import Dimensions
 from ngio.ome_zarr_meta.ngio_specs import DefaultSpaceUnit, PixelSize, SpaceUnits
 from ngio.utils import NgioValueError
